Update only updateable parameters with enif
PS: This is theoretically wrong, as parameters
may affect responses even if they are not updated.
Example of this:
Suppose we have one scalar parameter + 200 design
matrix parameters, and the response is mainly
affected by the design parameters. Now EnIF will
overlook this and give a poor update.

We should aim to first disregard categoricals,
but eventually also somehow include them.